Rebase dart-ffi to master (#32)
* Add abstract client trait.

* Move suri to client.

* Add generic utils.

* More stuff.

* Add generic cli.

* Fix build.

* Remove unused deps.

* Refactor test-cli to use cli-identity.

* Faucet works.
