+++ Cargo.toml
@@ -6,10 +6,13 @@ members = [
     "bin/runtime",
     "faucet/client",
     "faucet/pallet",
+    "identity/cli",
     "identity/client",
     "identity/pallet",
     "identity/utils",
     "identity/ffi",
     "keystore",
-
 ]
+
+[patch.crates-io]
+substrate-subxt = { git = "https://github.com/dvc94ch/substrate-subxt", branch = "unsigned" }
