+++ bin/cli/Cargo.toml
@@ -8,15 +8,10 @@ edition = "2018"
 async-std = { version = "=1.5.0", features = ["attributes"] }
 exitfailure = "0.5.1"
 clap = "3.0.0-beta.1"
-client-identity = { path = "../../identity/client" }
+cli-identity = { path = "../../identity/cli" }
 dirs2 = "3.0.1"
 env_logger = "0.7.1"
 ipfs-embed = "0.1.1"
 keybase-keystore = { path = "../../keystore" }
-qr2term = "0.2.1"
-rpassword = "4.0.5"
-sp-keyring = "2.0.0-rc3"
 substrate-subxt = "0.9.0"
 test-client = { path = "../client" }
-textwrap = { version = "0.11.0", features = ["term_size"] }
-thiserror = "1.0.20"
