+++ faucet/client/src/lib.rs
@@ -1,13 +1,20 @@
-use codec::Encode;
+use codec::{Decode, Encode};
+use substrate_subxt::balances::{Balances, BalancesEventsDecoder};
 use substrate_subxt::system::{System, SystemEventsDecoder};
-use substrate_subxt::{module, Call};
+use substrate_subxt::{module, Call, Event};
 
 #[module]
-pub trait Faucet: System {}
+pub trait Faucet: Balances + System {}
 
 #[derive(Call, Clone, Debug, Eq, Encode, PartialEq)]
 pub struct MintCall<'a, T: Faucet> {
-    account: &'a <T as System>::AccountId,
+    pub account: &'a <T as System>::AccountId,
+}
+
+#[derive(Clone, Debug, Decode, Eq, Event, PartialEq)]
+pub struct MintedEvent<T: Faucet> {
+    pub account: <T as System>::AccountId,
+    pub amount: <T as Balances>::Balance,
 }
 
 #[cfg(test)]
@@ -15,12 +22,12 @@ mod tests {
     use sp_core::sr25519::Pair;
     use sp_core::Pair as _;
     use substrate_subxt::{sp_core, ClientBuilder, PairSigner, Signer};
-    use test_client::faucet::MintCallExt;
+    use test_client::faucet::{MintCall, MintedEventExt};
+    use test_client::identity::Identity;
     use test_client::mock::test_node;
     use test_client::Runtime;
 
     #[async_std::test]
-    #[ignore]
     async fn test_mint() {
         let (node, _) = test_node();
         let client = ClientBuilder::<Runtime>::new()
@@ -28,12 +35,23 @@ mod tests {
             .build()
             .await
             .unwrap();
-        //let hans = PairSigner::<Runtime, _>::new(Pair::generate().0);
-        let hans = PairSigner::<Runtime, _>::new(test_client::mock::AccountKeyring::Alice.pair());
+        let hans = PairSigner::<Runtime, _>::new(Pair::generate().0);
+
+        let call = MintCall {
+            account: hans.account_id(),
+        };
+        let unsigned = client
+            .create_unsigned(call, hans.account_id(), None)
+            .await
+            .unwrap();
+        let mut decoder = client.events_decoder::<MintCall<Runtime>>();
+        decoder.register_type_size::<<Runtime as Identity>::Uid>("Uid");
 
         client
-            .mint_and_watch(&hans, hans.account_id())
+            .submit_and_watch_extrinsic(unsigned, decoder)
             .await
+            .unwrap()
+            .minted()
             .unwrap();
     }
 }
