+++ identity/cli/src/error.rs
@@ -5,22 +5,22 @@ pub enum Error {
     #[error(transparent)]
     Identity(#[from] client_identity::Error),
     #[error(transparent)]
-    Ipfs(#[from] ipfs_embed::Error),
-    #[error(transparent)]
     Subxt(#[from] substrate_subxt::Error),
     #[error(transparent)]
     Io(#[from] async_std::io::Error),
     #[error(transparent)]
-    Keystore(#[from] keybase_keystore::Error),
+    Keystore(#[from] keystore::Error),
     #[error(transparent)]
     Qr(#[from] qr2term::QrError),
 
     #[error("Failed to find config dir. Use `--path` to supply a suitable directory.")]
     ConfigDirNotFound,
-    #[error("Invalid suri encoded key pair.")]
-    InvalidSuri,
-    #[error("Invalid ss58 encoded account id.")]
-    InvalidSs58,
+    #[error(transparent)]
+    InvalidSuri(#[from] client_identity::InvalidSuri),
+    #[error(transparent)]
+    InvalidSs58(#[from] client_identity::InvalidSs58),
+    #[error(transparent)]
+    InvalidService(#[from] client_identity::ServiceParseError),
     #[error("Failed to decode transfer event.")]
     TransferEventDecode,
     #[error("Failed to find transfer event.")]
@@ -36,3 +36,5 @@ pub enum Error {
     #[error("Invalid paperkey.")]
     InvalidMnemonic,
 }
+
+pub type Result<T> = core::result::Result<T, Error>;
