+++ identity/client/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 
 [dependencies]
 async-std = { version = "=1.5.0", features = ["unstable"] }
+async-trait = "0.1.36"
 codec = { version = "1.3.0", package = "parity-scale-codec" }
 frame-support = "2.0.0-rc3"
 ipld-block-builder = "0.3.0"
@@ -17,7 +18,6 @@ serde_json = "1.0.55"
 substrate-subxt = { version = "0.9.0", features = ["client"] }
 ureq = { version = "^1", default-features = false, features = ["tls", "json"] }
 thiserror = "1.0.20"
-utils-identity = { version = "0.1.0", path = "../utils" }
 
 [dev-dependencies]
 async-std = { version = "=1.5.0", features = ["attributes"] }
