+++ identity/client/src/subxt.rs
@@ -1,6 +1,7 @@
 //! Subxt calls.
 use codec::{Decode, Encode, FullCodec};
 use core::convert::TryInto;
+use core::fmt::Display;
 use frame_support::Parameter;
 use libipld::cid::{Cid, Error as CidError};
 use std::str::FromStr;
@@ -10,7 +11,7 @@ use substrate_subxt::{module, Call, Event, Store};
 
 #[module]
 pub trait Identity: System {
-    type Uid: Parameter + Member + Copy + Default + CheckedAdd + Into<u64> + FromStr;
+    type Uid: Parameter + Member + Copy + Default + CheckedAdd + Into<u64> + FromStr + Display;
 
     type Cid: Parameter + Member + Default + From<Cid> + TryInto<Cid, Error = CidError>;
